backend/requirements.txt
fastapi>=0.118.0
uvicorn[standard]>=0.23.2
pydantic>=2.6.2
motor>=3.7.1
python-dotenv
openai>=1.33.0
httpx>=0.28.1
faiss-cpu
chromadb 
redis 
sentence-transformers